refactor: internal/healthmanager
