+++ internal/healthmanager/blockchain_health_manager_test.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/stretchr/testify/suite"
 
-	"github.com/status-im/status-go/healthmanager/rpcstatus"
+	"github.com/status-im/status-go/internal/healthmanager/rpcstatus"
 )
 
 type BlockchainHealthManagerSuite struct {
