+++ rpc/events.go
@@ -1,7 +1,7 @@
 package rpc
 
 import (
-	"github.com/status-im/status-go/healthmanager"
+	"github.com/status-im/status-go/internal/healthmanager"
 )
 
 type EventBlockchainHealthChanged struct {
