+++ services/wallet/api.go
@@ -20,8 +20,8 @@ import (
 	"github.com/status-im/status-go/accounts-management/generator"
 	"github.com/status-im/status-go/crypto"
 	"github.com/status-im/status-go/crypto/types"
-	"github.com/status-im/status-go/healthmanager"
 	abi_spec "github.com/status-im/status-go/internal/abi-spec"
+	"github.com/status-im/status-go/internal/healthmanager"
 	"github.com/status-im/status-go/logutils"
 	"github.com/status-im/status-go/params"
 	"github.com/status-im/status-go/rpc/network"
